requirements.txt
pandas
plotly.express
streamlit
json
requests
unicodedata

